Encapsulate x, y, width, height, top_left and bottom_right. Use size_type and point_type for storage, reduce copy operations.

+++ gui/core/sources/guidefs.h
@@ -96,8 +96,8 @@ namespace gui {
 
     typedef XPoint point_type;
     typedef struct {
-        unsigned int cx;
-        unsigned int cy;
+        unsigned short cx;
+        unsigned short cy;
     } size_type;
     typedef XRectangle rectangle_type;
     typedef XFontStruct* font_type;
